refactor limited maps creators, moving it to separate LimitedMapBackingMaps
