+++ infrastructure/collections/src/main/java/tech/pegasys/teku/infrastructure/collections/NonSynchronizedLimitedMap.java
@@ -17,7 +17,7 @@ package tech.pegasys.teku.infrastructure.collections;
 final class NonSynchronizedLimitedMap<K, V> extends AbstractLimitedMap<K, V> {
 
   public NonSynchronizedLimitedMap(final int maxSize) {
-    super(createLimitedMap(maxSize, true), maxSize);
+    super(LimitedMapBackingMaps.createLimitedMap(maxSize, true), maxSize);
   }
 
   @Override
